AI Agent/freelance_scout/requirements.txt
fastapi
uvicorn
crewai
crewai-tools
pydantic
beautifulsoup4
python-dotenv
